Convert to Maven Multimodule
Changes:
  * Move to a maven multimodule system. This allows dependencies to be specifically targetted
  * Abstract out the BetonQuest version interface
  * Fix a failed conflict in plugin.yml from upstream
