+++ nuBeton_jobsReborn-api/pom.xml
@@ -1,0 +1,28 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <parent>
+        <artifactId>nuBeton_JobsReborn-parent</artifactId>
+        <groupId>net.livecar.nuttyworks.nubeton_jobsreborn</groupId>
+        <version>1.13-SNAPSHOT</version>
+    </parent>
+    <modelVersion>4.0.0</modelVersion>
+
+    <artifactId>nuBeton_jobsReborn-api</artifactId>
+
+    <dependencies>
+
+        <!-- Earliest Release we want to support -->
+        <dependency>
+            <groupId>org.spigotmc</groupId>
+            <artifactId>spigot-api</artifactId>
+            <!--<version>1.13-R0.1-SNAPSHOT</version>-->
+            <!--<version>1.12.2-R0.1-SNAPSHOT</version>-->
+            <version>1.8.8-R0.1-SNAPSHOT</version>
+            <scope>provided</scope>
+        </dependency>
+    </dependencies>
+
+
+</project>
