+++ nuBeton_jobsReborn-api/src/main/java/net/livecar/NuttyWorks/BetonQuestPlugin.java
@@ -1,0 +1,16 @@
+package net.livecar.NuttyWorks;
+
+import org.bukkit.plugin.java.JavaPlugin;
+
+public abstract class BetonQuestPlugin
+{
+    private static JavaPlugin plugin;
+
+    public BetonQuestPlugin(JavaPlugin sourcePlugin) {
+        plugin = sourcePlugin;
+    }
+
+    public static JavaPlugin getPlugin() {
+        return plugin;
+    }
+}
