+++ nuBeton_jobsReborn-betonquest-v1_8/pom.xml
@@ -1,0 +1,55 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <parent>
+        <artifactId>nuBeton_JobsReborn-parent</artifactId>
+        <groupId>net.livecar.nuttyworks.nubeton_jobsreborn</groupId>
+        <version>1.13-SNAPSHOT</version>
+    </parent>
+    <modelVersion>4.0.0</modelVersion>
+
+    <artifactId>nuBeton_jobsReborn-betonquest-v1_8</artifactId>
+
+    <dependencies>
+        <!-- Local -->
+        <dependency>
+            <groupId>${project.parent.groupId}</groupId>
+            <artifactId>nuBeton_jobsReborn-api</artifactId>
+            <version>${project.parent.version}</version>
+            <scope>provided</scope>
+        </dependency>
+
+        <!-- Earliest Release we want to support -->
+        <dependency>
+            <groupId>org.spigotmc</groupId>
+            <artifactId>spigot-api</artifactId>
+            <!--<version>1.13-R0.1-SNAPSHOT</version>-->
+            <!--<version>1.12.2-R0.1-SNAPSHOT</version>-->
+            <version>1.8.8-R0.1-SNAPSHOT</version>
+            <scope>provided</scope>
+        </dependency>
+
+        <!-- JobsReborn -->
+        <dependency>
+            <groupId>com.gamingmesh.jobs</groupId>
+            <artifactId>Jobs</artifactId>
+            <version>4.8.0</version>
+            <scope>provided</scope>
+        </dependency>
+
+        <!-- BetonQuest -->
+        <dependency>
+            <groupId>pl.betoncraft.betonquest</groupId>
+            <artifactId>BetonQuest</artifactId>
+            <version>1.8.5</version>
+            <scope>provided</scope>
+        </dependency>
+
+
+
+
+    </dependencies>
+
+
+</project>
