+++ nuBeton_jobsReborn-betonquest-v1_8/src/main/java/net/livecar/NuttyWorks/nuBeton_JobsReborn_V1_8/Objectives/Objective_JoinJob.java
@@ -1,25 +1,24 @@
 package net.livecar.NuttyWorks.nuBeton_JobsReborn_V1_8.Objectives;
 
-import net.livecar.NuttyWorks.nuBeton_JobsReborn.BetonJobsReborn;
-
+import net.livecar.NuttyWorks.nuBeton_JobsReborn_V1_8.BetonJobsReborn_V1_8;
 import org.bukkit.Bukkit;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.HandlerList;
 import org.bukkit.event.Listener;
 
 import com.gamingmesh.jobs.Jobs;
-import com.gamingmesh.jobs.api.JobsLevelUpEvent;
+import com.gamingmesh.jobs.api.JobsJoinEvent;
 import com.gamingmesh.jobs.container.Job;
 
 import pl.betoncraft.betonquest.InstructionParseException;
 import pl.betoncraft.betonquest.api.Objective;
 import pl.betoncraft.betonquest.utils.PlayerConverter;
 
-public class Objective_LevelUpEvent extends Objective implements Listener 
+public class Objective_JoinJob extends Objective implements Listener 
 {
 	private final String sJobName;
 	
-	public Objective_LevelUpEvent(String packName, String label, String instructions) throws InstructionParseException 
+	public Objective_JoinJob(String packName, String label, String instructions) throws InstructionParseException 
 	{
         super(packName, label, instructions);
         template = ObjectiveData.class;
@@ -39,9 +38,9 @@ public class Objective_LevelUpEvent extends Objective implements Listener
     }
 	
 	@EventHandler
-    public void onJobsLevelUpEvent(JobsLevelUpEvent event) 
+    public void onJobsJoinEvent(JobsJoinEvent event) 
 	{
-		if (event.getJobName().equalsIgnoreCase(this.sJobName))
+		if (event.getJob().getName().equalsIgnoreCase(this.sJobName))
 		{
 			String playerID = PlayerConverter.getID(event.getPlayer().getPlayer().getPlayer());
             if (containsPlayer(playerID) && checkConditions(playerID)) {
@@ -52,7 +51,7 @@ public class Objective_LevelUpEvent extends Objective implements Listener
     
     @Override
     public void start() {
-        Bukkit.getPluginManager().registerEvents(this, BetonJobsReborn.Instance);
+        Bukkit.getPluginManager().registerEvents(this, BetonJobsReborn_V1_8.getPlugin());
     }
 
     @Override
