+++ nuBeton_jobsReborn/src/main/java/net/livecar/NuttyWorks/nuBeton_JobsReborn_V1_9/BetonJobsReborn_V1_9.java
@@ -1,25 +1,30 @@
-package net.livecar.NuttyWorks.nuBeton_JobsReborn_V1_8;
+package net.livecar.NuttyWorks.nuBeton_JobsReborn_V1_9;
 
 import java.util.logging.Level;
 
+import net.livecar.NuttyWorks.BetonQuestPlugin;
 import org.bukkit.Bukkit;
 
-import net.livecar.NuttyWorks.nuBeton_JobsReborn_V1_8.Conditions.*;
-import net.livecar.NuttyWorks.nuBeton_JobsReborn_V1_8.Events.*;
-import net.livecar.NuttyWorks.nuBeton_JobsReborn_V1_8.Objectives.*;
+import net.livecar.NuttyWorks.nuBeton_JobsReborn_V1_9.Conditions.*;
+import net.livecar.NuttyWorks.nuBeton_JobsReborn_V1_9.Events.*;
+import net.livecar.NuttyWorks.nuBeton_JobsReborn_V1_9.Objectives.*;
+
+import org.bukkit.plugin.java.JavaPlugin;
 import pl.betoncraft.betonquest.BetonQuest;
 
-public class BetonJobsReborn_V1_8
+public class BetonJobsReborn_V1_9 extends BetonQuestPlugin
 {
-	public BetonJobsReborn_V1_8()
+	public BetonJobsReborn_V1_9(JavaPlugin plugin)
 	{
+	    super(plugin);
+
 		//Register conditions with beton
 		BetonQuest.getInstance().registerConditions("nujobs_canlevel", Condition_CanLevel.class);
 		BetonQuest.getInstance().registerConditions("nujobs_hasjob", Condition_HasJob.class);
 		BetonQuest.getInstance().registerConditions("nujobs_jobfull", Condition_JobFull.class);
 		BetonQuest.getInstance().registerConditions("nujobs_joblevel", Condition_JobLevel.class);
 		Bukkit.getServer().getLogger().log(Level.INFO, "Registered Conditions [nujobs_canlevel,nujobs_hasjob,nujobs_jobfull,nujobs_joblevel]");
-		
+	    
 		//register events
 		BetonQuest.getInstance().registerEvents("nujobs_addexp", Event_AddExp.class);
 		BetonQuest.getInstance().registerEvents("nujobs_addlevel", Event_AddLevel.class);
@@ -35,6 +40,5 @@ public class BetonJobsReborn_V1_8
 		BetonQuest.getInstance().registerObjectives("nujobs_levelup", Objective_LevelUpEvent.class);
 		BetonQuest.getInstance().registerObjectives("nujobs_payment", Objective_PaymentEvent.class);
 		Bukkit.getServer().getLogger().log(Level.INFO, "Registered Objectives [nujobs_joinjob,nujobs_leavejob,nujobs_levelup,nujobs_payment]");
-		
 	}
 }
