+++ nuBeton_jobsReborn/src/main/java/net/livecar/NuttyWorks/nuBeton_JobsReborn_V1_9/Conditions/Condition_HasJob.java
