+++ nuBeton_jobsReborn/src/main/resources/plugin.yml
@@ -1,0 +1,6 @@
+name: NuBeton_JobsReborn
+main: net.livecar.NuttyWorks.nuBeton_JobsReborn.BetonJobsReborn
+version: 1.12
+description: Extends BetonQuest to interface with Jobs reborn
+authors: [Nutty101]
+depend: [Jobs,BetonQuest]
