--- plugin.yml
@@ -1,15 +1,0 @@
-<<<<<<< HEAD
-name: NuBeton_JobsReborn
-main: net.livecar.NuttyWorks.nuBeton_JobsReborn.BetonJobsReborn
-version: 1.12
-description: Extends BetonQuest to interface with Jobs reborn
-authors: [Nutty101]
-depend: [Jobs,BetonQuest]
-=======
-name: NuBeton_JobsReborn
-main: net.livecar.NuttyWorks.nuBeton_JobsReborn.BetonJobsReborn
-version: 1.8
-description: Extends BetonQuest to interface with Jobs reborn
-authors: [Nutty101]
-depend: [Jobs,BetonQuest]
->>>>>>> origin/master
