--- src/net/livecar/NuttyWorks/nuBeton_JobsReborn/BetonJobsReborn_Interface.java
@@ -1,6 +1,0 @@
-package net.livecar.NuttyWorks.nuBeton_JobsReborn;
-
-public interface BetonJobsReborn_Interface 
-{
-
-}
